--- src/main/java/subway/line/LineRequest.java
@@ -1,57 +1,0 @@
-package subway.line;
-
-import subway.section.Section;
-
-public class LineRequest {
-    private static final int INITIAL_DISTANCE = 0;
-
-    private String name;
-    private String color;
-    private Long upStationId;
-    private Long downStationId;
-    private int distance;
-    private int extraFare;
-
-    public LineRequest() {
-    }
-
-    public LineRequest(String name, String color, Long upStationId, Long downStationId, int distance) {
-        this.name = name;
-        this.color = color;
-        this.upStationId = upStationId;
-        this.downStationId = downStationId;
-        this.distance = distance;
-    }
-
-    public Section toFirstSection(Long lineId) {
-        return new Section(lineId, upStationId, INITIAL_DISTANCE);
-    }
-
-    public Section toSection(Long lineId) {
-        return new Section(lineId, downStationId, distance);
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public Long getUpStationId() {
-        return upStationId;
-    }
-
-    public Long getDownStationId() {
-        return downStationId;
-    }
-
-    public int getDistance() {
-        return distance;
-    }
-
-    public int getExtraFare() {
-        return extraFare;
-    }
-}
