+++ src/main/java/subway/line/controller/LineController.java
@@ -1,0 +1,52 @@
+package subway.line.controller;
+
+import org.springframework.http.MediaType;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import subway.line.dto.LineRequest;
+import subway.line.dto.LineResponse;
+import subway.line.service.LineService;
+
+import java.net.URI;
+import java.util.List;
+
+@RestController
+@RequestMapping("/lines")
+public class LineController {
+    private final LineService lineService;
+
+    public LineController(LineService lineService) {
+        this.lineService = lineService;
+    }
+
+    @PostMapping()
+    public ResponseEntity<LineResponse> createLine(@RequestBody LineRequest lineRequest) {
+        LineResponse lineResponse = lineService.createLine(lineRequest);
+        return ResponseEntity.created(URI.create("/lines/" + lineResponse.getId())).body(lineResponse);
+    }
+
+    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
+    public ResponseEntity<List<LineResponse>> showLines() {
+        return ResponseEntity.ok().body(lineService.getLineResponses());
+    }
+
+    @DeleteMapping("/{lineId}")
+    public ResponseEntity deleteLine(@PathVariable Long lineId) {
+        lineService.deleteLineById(lineId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping(value = "/{lineId}")
+    public ResponseEntity<LineResponse> showLineById(@PathVariable Long lineId) {
+        LineResponse lineResponse = lineService.showLineById(lineId);
+        return ResponseEntity.ok().body(lineResponse);
+    }
+
+    @PutMapping(value = "/{lineId}")
+    public ResponseEntity putLine(@PathVariable Long lineId, @RequestBody LineRequest lineRequest) {
+        lineService.updateLineByLineId(lineId,lineRequest);
+        return ResponseEntity.ok().build();
+    }
+
+}
+
