--- src/main/java/subway/section/SectionController.java
@@ -1,29 +1,0 @@
-package subway.section;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import subway.line.LineResponse;
-
-@RestController
-@RequestMapping("/lines/{lineId}/sections")
-public class SectionController {
-
-    private final SectionService sectionService;
-
-    public SectionController(SectionService sectionService) {
-        this.sectionService = sectionService;
-    }
-
-    @PostMapping()
-    public ResponseEntity<LineResponse> createSection(@PathVariable Long lineId, @RequestBody SectionRequest sectionRequest) {
-        sectionService.createSection(lineId,sectionRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @DeleteMapping()
-    public ResponseEntity<LineResponse> deleteSection(@PathVariable Long lineId, @RequestParam Long stationId) {
-        sectionService.deleteSection(lineId, stationId);
-        return ResponseEntity.ok().build();
-    }
-
-}
