--- src/main/java/subway/station/StationRequest.java
@@ -1,18 +1,0 @@
-package subway.station;
-
-public class StationRequest {
-    private String name;
-
-    public StationRequest() {
-
-    }
-
-    public StationRequest(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-}
