--- src/main/java/subway/station/StationResponse.java
@@ -1,24 +1,0 @@
-package subway.station;
-
-public class StationResponse {
-    private Long id;
-    private String name;
-
-    public StationResponse(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public static StationResponse from(Station station) {
-        return new StationResponse(station.getId(), station.getName());
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-}
