+++ src/main/java/subway/station/domain/Station.java
@@ -1,0 +1,28 @@
+package subway.station.domain;
+
+public class Station {
+    private Long id;
+    private String name;
+
+    public Station() {
+    }
+
+    public Station(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public Station(String name) {
+        this.name = name;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+}
+
