+++ src/main/java/subway/station/dto/StationRequest.java
@@ -1,0 +1,18 @@
+package subway.station.dto;
+
+public class StationRequest {
+    private String name;
+
+    public StationRequest() {
+
+    }
+
+    public StationRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+}
